Added additional vector accessors + simplified lib
This commit adds some accessors known in shader languages. In addition
I removed the overcomplicated functions mmx_memset and mmx_memcpy and replaced
it with a simpler implementation. If something more optimized is needed
it is still possible to override the function by definining makro
MMX_MEMSET and MMX_MEMCOPY. I also removed the need for <stdint.h> and
the library can in theory outside math functions be used without the
standard library.