Fix lexer_read_on_line returning empty tokens at EOF
At EOF, lexer_read() returns 0 and an empty token.
While this is supposedly taken care of by the if statement it is called
in, the statement falls through and reaches the following
    if (!tok.line_crossed)
which may evaluate to true (UB due to line_crossed being undefined)
and consequently return empty tokens indefinitely.